# public/openapi.yaml
openapi: 3.1.0
info:
  title: ScoutEdge B2B Data API
  version: "1.0"
  description: |
    Real-time World Cup 2026 intelligence — matches, predictions, signals,
    odds, value bets, team chemistry, and player intel.

    ## Authentication

    All requests require an API key passed via the `X-API-Key` header.

    ```
    X-API-Key: se_live_abc123...
    ```

    Keys are provisioned per-organization. Contact sales@scoutedge.com to
    obtain a key or upgrade your tier.

    ## Tiers

    | Tier | Endpoints | Rate Limit |
    |------|-----------|------------|
    | **Basic** | matches, standings, predictions, teams | Per plan |
    | **Advanced** | All Basic + odds, signals, value-bets, players, chemistry | Per plan |
    | **Event** | All Advanced + real-time websocket (coming soon) | Per plan |
    | **White Label** | Full platform embed | Custom |

    ## Rate Limiting

    Every response includes rate-limit headers:

    | Header | Description |
    |--------|-------------|
    | `X-RateLimit-Limit` | Max requests in current window |
    | `X-RateLimit-Remaining` | Requests remaining |
    | `X-RateLimit-Reset` | ISO 8601 timestamp when window resets |
    | `Retry-After` | Seconds to wait (only on 429) |

    Both per-minute and per-month limits apply. When either is exceeded the
    API responds with `429 Too Many Requests`.

    > **v1.0 note:** Per-minute rate limits are currently enforced **per serverless instance**.
...

    ## Response Envelope

...

    ```json
    {
...

...

...
servers:
...
security:
  - ApiKeyAuth: []
...
paths:
...
  /predictions:
    get:
      operationId: listPredictions
      summary: List predictions
      description: Returns AI-generated match predictions with win probabilities.
      tags: [Predictions]
      x-codeSamples:
        - lang: cURL
          source: |
            curl -H "X-API-Key: se_live_YOUR_KEY" \
              "https://scoutedge.com/api/v1/predictions?date=2026-06-15"
        - lang: Python
          source: |
            import httpx

            resp = httpx.get(
                "https://scoutedge.com/api/v1/predictions",
                headers={"X-API-Key": "se_live_YOUR_KEY"},
                params={"date": "2026-06-15"},
            )
            predictions = resp.json()["data"]
        - lang: JavaScript
          source: |
            const res = await fetch(
              "https://scoutedge.com/api/v1/predictions?date=2026-06-15",
              { headers: { "X-API-Key": "se_live_YOUR_KEY" } }
            );
            const { data: predictions } = await res.json();
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: date
          in: query
          schema: { type: string, format: date, example: "2026-06-15" }
          description: Filter by match date (YYYY-MM-DD)
        - name: team_id
          in: query
          schema: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440001" }
          description: Filter by team ID
      responses:
        "200":
          description: Paginated prediction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionListResponse'
        "401": { $ref: '#/components/responses/Unauthorized' }
        "429": { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Your ScoutEdge API key (e.g. `se_live_abc123...`)
  parameters:
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Page number (1-based)
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Results per page (max 100)
...
  schemas:
    Meta:
      type: object
      properties:
        total: { type: integer, example: 48 }
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
      required: [total, page, limit]
    ErrorEnvelope:
      type: object
      properties:
        status: { type: string, const: "error" }
        data: { type: "null" }
        error: { type: string, example: "Invalid or expired API key" }
      required: [status, data, error]
...
    Prediction:
      type: object
      properties:
        id: { type: string, format: uuid }
        match_id: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440000" }
        home_team_name: { type: string, example: "Brazil" }
        away_team_name: { type: string, example: "France" }
        home_win_prob: { type: number, format: float, example: 0.42 }
        draw_prob: { type: number, format: float, example: 0.28 }
        away_win_prob: { type: number, format: float, example: 0.30 }
        explanation_json: { type: object }
        confidence_interval: { type: number, format: float, example: 0.055 }
        model_version: { type: string, example: "elo-poisson-v2" }
        created_at: { type: string, format: date-time }
        kickoff_utc: { type: string, format: date-time, nullable: true }
    PredictionListResponse:
      type: object
      properties:
        status: { type: string, const: "ok" }
        data: { type: array, items: { $ref: '#/components/schemas/Prediction' } }
        meta: { $ref: '#/components/schemas/Meta' }
      required: [status, data]
...
  responses:
...
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: error
            data: null
            error: "Missing X-API-Key header"
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema: { type: integer }
          description: Max requests in current window
        X-RateLimit-Remaining:
          schema: { type: integer }
          description: Requests remaining (0 when exceeded)
        X-RateLimit-Reset:
          schema: { type: string, format: date-time }
          description: When the rate limit window resets
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            status: error
            data: null
            error: "Rate limit exceeded"
...